Create database FRMs to be used within akka stream (#4)
# Why
Soundoff pages should parse not only the ratings, but also the album itself so that the album can be inserted into the database as well as the join table populated.

# What
* Brought in strongtypes
* First draft of akka stream with dummy data
* Added all FRMs

+++ src/main/scala/migrations/CreateBands.scala
@@ -13,7 +13,7 @@ object CreateBands extends Migration {
     .addIndexes(_.nameIndex)
 
   class Bands(tag: Tag) extends Table[Unit](tag, "bands") {
-    def id = column[Long]("id", O.PrimaryKey)
+    def id = column[Int]("id", O.PrimaryKey)
 
     def name = column[String]("name")
